Fix update pref for member of only one program
