+++ frontend/src/components/profile/functions/index.ts
@@ -1,3 +1,2 @@
 export * from './ProfileFunctionEdit';
 export * from './ProfileFunctions';
-export * from './MemberProfileFunctions';
