+++ frontend/src/components/profile/sections-roles/index.ts
@@ -2,4 +2,5 @@ export * from './ProfileEditListSection';
 export * from './ProfileEditRoles';
 export * from './SectionsAndRoles';
 export * from './MemberItemList';
-export * from './MemberProfileEditRoles';
+export * from './MemberProfileEditPreferences';
+export * from './MemberProfileEditRoles';
